Introducing service mocking feature
Mocking features can help varlink users to simplify varlink client
implementation and testing like python-podman by example.

This feature is useful for unit testing to quickly setup a varlink
server based on a mock class who will return user defined values.

This feature avoid to setup an env with the server part to test our
clients, and user can be focused on feature development and also merely unit
test them with only a pur python environment.

The mock module introduce a function decorator and user can use it
to start the mock server and surround specific tests with it.

This feature only work with python3+, and is not compatible with python2
due to the usage of annotations.

These changes also introduce tox py27 target to handle the specific use case
of the mocking module and related tests that we need to be
ignored/excluded.

If users try to import this module under python2 then it will raise an
ImportException to avoid SyntaxError related to annotations.