README.md: Update python invocation in examples
The code examples in the README.md file were inconsistent - some of them
were invoking the code with `python`, which others were using `python3`
explicitly.

This commit updates the examples in this file to always invoke the code
with `python3`, to ensure they are also working in systems where
`python` is not available.