package: Drop setuptools_scm_git_archive requirement
The tool of setuptools_scm_git_archive is deprecated. The logic of this
package was introduced directly to the setuptools_scm package, since
version 7.0.0 of this package.

This commit removes the unnecessary requirement to have
setuptools_scm_git_archive, and instead, ensures that the setuptools_scm
version used is recent enough.