pyproject: add ruff and mypy testenvs
Add ruff an mypy test envs for tox to pyproject.toml. Due to differences
in the tox TOML config format the mypy envs are set up with py3- and
py39- prefixes for default python3 and python3.9 testing respectively.
In my experience testing the oldest supported and current versions is
generally enough for mypy.

Signed-off-by: John Mulligan <jmulligan@redhat.com>