setup: remove setup.cfg in favour of pyproject.toml
The option zip_safe is dropped is dropped, since it is considered obselete by
setuptools and no longer supposed to do anything [1]. The option
bdist_wheel.universal is dropped as well, since this flag used to show that a
wheel was compatible with both Python 2 and 3 [2] and support for Python 2 has
been dropped.

More project urls are added, since pyproject.toml uses a richer section there
instead of the previous single url field.

The license classifier is switched to the SPDX identifier for the Apache 2.0
license and a license-files array linking to the respective file is added. The
license classifier is removed, because at least some build backends refuse to
build the package with both it and other license metadata at the same time in
the file.

[1] https://setuptools.pypa.io/en/latest/deprecated/zip_safe.html
[2] https://wheel.readthedocs.io/en/stable/quickstart.html

debug