client: drop namespaced from VarlinkError.new in _next_varlink_message
The relevant code is in essence the following

    decoded = json.loads(message)
    if "error" in decoded and decoded["error"] is not None:
        raise VarlinkError.new(decoded)

where message is at first an incoming Varlink message that is string-encoded
JSON. This means at this point the thing being handed to VarlinkError.new is a
dictionary. The namespaced keyword being passed in before, would tell
VarlinkError to handle the incoming thing as a SimpleNamespace, thus leading to
breakage.

An alternative at this point might be to instead add an object hook if
namespaced is true, but since the instantiation of the error object will
directly normalise this to a dictionary and the relevant methods on the error
have their own namespace options, this seems gratuitous.